typeahead-options renamed to typeaheadOptions, removing some proptype updates that are out of date
