+++ assets/src/dashboard/components/index.js
@@ -20,7 +20,7 @@ export { default as Button } from './button';
 export { default as PopoverMenu } from './popover-menu';
 export { default as Dropdown } from './dropdown';
 export { default as TypeaheadInput } from './typeaheadInput';
-export { default as TypeaheadOptions } from './typeahead-options';
+export { default as TypeaheadOptions } from './typeaheadOptions';
 export { Pill, FloatingTab } from './pill';
 export { ViewHeader } from './typography';
 export { default as CardGrid, StoryGrid } from './cardGrid';
